Add Bysykler and SmartBike system
The get_all() method takes TOO long, as it updates
all the stations. Nevertheless, it's assumable for
Olso (105 stations), and the SmartBike API does
not return any useful information in the listing
url (only id). It even returns invalid stations...
which we do not know are invalid until we try
to update info on them.. Shame!