fix license mismatch
through the years the library has been both licensed as LGPL and AGPL,
as a result of confusion on which one. Scattered around the files there
are both AGPL and LGPL notices. I have decided to go with the one
covering the majority of the files and also the one I feel closer to the
project: AGPL